Task OLTSWITCH-1343 Upgrade Switch Driver SDK from 6.5.18 to 6.5.20:  - Applied sdk-all-6.5.20_s6_switchdrvr.patch (adaptions for switchdrvr)
svn path=/trunk/; revision=7736

+++ lib/broadcom-sdk-xgs/sdk-all-switchdrvr-6.5.20/broadcom/esw_sdk/include/appl/diag/dpp/utils_eeprom.h
@@ -638,7 +638,7 @@ typedef struct
      * Number identifier of task (where reset exception
      * has occurred).
      */
-  int            task_id ;
+  unsigned long long task_id ;
     /*
      * Free text description of exception. Null terminated
      * string.
@@ -831,7 +831,7 @@ typedef struct
      * Number identifier of task (where reset exception
      * has occurred).
      */
-  int            task_id ;
+  unsigned long long task_id ;
     /*
      * Free text description of exception. Null terminated
      * string.
